+++ src/main/java/racingcar/RacingService.java
@@ -1,6 +1,7 @@
 package racingcar;
 
 import java.util.List;
+import racingcar.domain.Car;
 
 public class RacingService {
     public void move(List<Car> carList) {
