+++ src/main/java/racingcar/message/ViewMessage.java
@@ -1,6 +1,6 @@
-package racingcar;
+package racingcar.message;
 
-public interface Message {
+public interface ViewMessage {
     String INPUT_CAR_NAME = "경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)";
     String INPUT_TRIAL_COUNT = "시도할 회수는 몇회인가요?";
 }
